requirements.txt
numpy==2.1.3
pandas==2.2.2
scikit-learn==1.5.1
xgboost==2.1.1
streamlit==1.38.0
requests==2.32.3
python-dotenv==1.0.1
